requirements.txt
# OpenAI CLIP
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
open_clip_torch
onnx
onnxruntime
onnxscript

# FastAPI
aiohttp
fastapi[standard-no-fastapi-cloud-cli]
fastapi-injectable

# AWS Lambda
mangum
